fix nAliqISS type

+++ src/OmieSharp/Models/CadastroServicoImpostos.cs
@@ -5,41 +5,41 @@ namespace OmieSharp.Models
 {
     public class CadastroServicoImpostos
     {
-        public int nAliqISS { get; set; }
+        public decimal nAliqISS { get; set; }
 
         [JsonConverter(typeof(BooleanSNJsonConverter))]
         public bool? cRetISS { get; set; }
 
-        public int nAliqPIS { get; set; }
+        public decimal nAliqPIS { get; set; }
 
         [JsonConverter(typeof(BooleanSNJsonConverter))]
         public bool? cRetPIS { get; set; }
 
-        public int nAliqCOFINS { get; set; }
+        public decimal nAliqCOFINS { get; set; }
 
         [JsonConverter(typeof(BooleanSNJsonConverter))]
         public bool? cRetCOFINS { get; set; }
 
-        public int nAliqCSLL { get; set; }
+        public decimal nAliqCSLL { get; set; }
 
         [JsonConverter(typeof(BooleanSNJsonConverter))]
         public bool? cRetCSLL { get; set; }
 
-        public int nAliqIR { get; set; }
+        public decimal nAliqIR { get; set; }
 
         [JsonConverter(typeof(BooleanSNJsonConverter))]
         public bool? cRetIR { get; set; }
 
-        public int nAliqINSS { get; set; }
+        public decimal nAliqINSS { get; set; }
 
         [JsonConverter(typeof(BooleanSNJsonConverter))]
         public bool? cRetINSS { get; set; }
 
-        public int nRedBaseINSS { get; set; }
+        public decimal nRedBaseINSS { get; set; }
 
-        public int nRedBaseCOFINS { get; set; }
+        public decimal nRedBaseCOFINS { get; set; }
 
-        public int nRedBasePIS { get; set; }
+        public decimal nRedBasePIS { get; set; }
         
         public bool lDeduzISS { get; set; }
     }
